Fix EZP-23302: Update Location fails if no change is performed with the update
@crevillo's PR #973 concludes that the exception should be commented, and wider design issue be solved separately, ref PDO::MYSQL_ATTR_FOUND_ROWS

While the exception is easy to provoke in a CLI script, I was unable to do so in a unit test. The added test passes before and after commenting the exception code :(

Edit: Uncommented the code again. The integration tests should be running on mysql too, so it should fail now. Set sail for fail!
Edit: Re-commented the code, since we see now that Travis fails without it, as it should.
Edit: Updated phpdoc about not throwing the exception anymore.