Fix EZP-23953: Legacy session events not triggered any more
> https://jira.ez.no/browse/EZP-23953

As of 5.3, authentication is completely managed through Symfony. One
side effect is that legacy session events are not triggered any more
(either via `ezpEvent` or `eZSession::triggerCallback()`), and thus
listeners won't be executed. This is because it is now entirely up to
Symfony to deal with the session, including session migration after
authentication.
**Concrete illustration of this regression :** When using the legacy
shop module, items in the shopping basket are lost after logging in.
In short, when the session is migrated after the user has been
authenticated, the basket table, storing each entry per session ID,
won't be updated at that time, resulting the loss of the basket for
current user.

This PR introduces proxies for session storage and session save handler.
They use legacy callbacks to trigger the appropriate events, just like
what is done in legacy session handlers.