Fix EZP-23971: If a user is removed while logged in, session for that user will throw an exception
> https://jira.ez.no/browse/EZP-23971

When in legacy mode, authentication is completely delegated to the
legacy kernel. However currently logged in user is also injected in the
Repository to make it usable with appropriate permissions.

Problem is that when the user is removed while a session for this user
is still open, Symfony stack still tries to inject it in the repository,
failing with a nasty exception.
This PR fixes the issue.

Cherry pick from LegacyBridge