Fix EZP-24017: Security token issue using legacy bridge
> https://jira.ez.no/browse/EZP-24017

Addresses the *Cannot set this token to trusted after instantiation*
exception when trying to authenticate both in frontend and backend.
This is a regression caused by fix for
[EZP-23953](https://jira.ez.no/browse/EZP-23953). Reason is that
`LegacySessionStorage` does not inherit from `NativeSessionStorage`, and
thus [session name and other session options from siteaccess
configuration are not passed to it any
more](|https://github.com/ezsystems/ezpublish-kernel/blob/master/eZ/Bundle/EzPublishCoreBundle/EventListener/SessionSetDynamicNameListener.php#L71-L89).

This patch fixes the issue by making `LegacySessionStorage` inherit from
`NativeSessionStorage`. This is only a workaround to strictly respect
its full interface, as `NativeSessionStorage` has public methods that
are not part of an interface. Those public methods have been
reimplemented and will act as proxies if `innerSessionStorage` is an
actual instance of `NativeSessionStorage`.