Fix EZP-23725: Wrong path_identification_string when editing in frontend siteaccess using PathPrefix and RootNode
> https://jira.ez.no/browse/EZP-23725

Alternative of ezsystems/ezpublish-legacy#1148,
ezsystems/ezpublish-legacy#1140 and
ezsystems/ezpublish-legacy#1145

Root cause of this issue is that `RootNode` is wrongly injected from
LegacyBundle when working in multisite, as it's has no purpose in that
regard.

Note that patch from
ezsystems/ezpublish-legacy#1140 needs to be
reverted (it will be once this PR is merged).

Tested and working as expected.