Fix EZP-24053: Preview locks content object, and cause version edit conflict
> https://jira.ez.no/browse/EZP-24053

This issue was due to an incomplete implementation of legacy website toolbar (handling preview was missing).
This PR ensures that:

* Content preview route is only used with Http GET (POST will be used by legacy to publish or go back to editing).
* One can easily check if preview is active and retrieve previewed content/location.

ContentPreviewHelper is now used to check if preview is active and to get previewed content/location
since as of Symfony 2.4, request attributes from the master request are not passed to sub-requests any more.
Furthermore, it's cleaner to be able to get this information through a dedicated service instead of request attributes.

(cherry picked from commit 7ebb250ce6eb35fbaeea7a76762f879ab70885c0)

Conflicts:
	eZ/Publish/Core/Helper/ContentPreviewHelper.php
	eZ/Publish/Core/Helper/Tests/ContentPreviewHelperTest.php
	eZ/Publish/Core/MVC/Symfony/Controller/Content/PreviewController.php